requirements.txt
# PPAForge AI - Requirements
# Python 3.9+ required

# Core ML/DL frameworks
torch>=2.0.0
torch-geometric>=2.3.0
torch-scatter>=2.1.0
torch-sparse>=0.6.17
torch-cluster>=1.6.1

# Reinforcement Learning
ray[rllib]>=2.5.0
gymnasium>=0.28.0
tensorflow>=2.12.0  # Required by RLlib

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=1.0.0
click>=8.1.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Web UI
streamlit>=1.28.0

# EDA integration (optional, may require manual installation)
# OpenROAD - Install separately from: https://github.com/The-OpenROAD-Project/OpenROAD
# gdspy>=1.6  # For DEF/GDS manipulation if needed

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0

# Logging and monitoring
tensorboard>=2.13.0
wandb>=0.15.0  # Optional: Weights & Biases for experiment tracking

# Jupyter (for notebooks)
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0

# Additional utilities
tqdm>=4.66.0
psutil>=5.9.0
colorlog>=6.7.0
